Implement several design changes
1) Move the title section inside sidebar, as there was area lying vacant
in sidebar, and also title section was consuming extra space of project section.
2) Remove dueTime feature of a todo. It was not needed at all, so just
got rid of it.
3) Earlier in the dialog box for creating a new todo, projects were displayed
using checkboxes. Replaced it with dropdown menu containing all the projects.
It solved the problem of user selecting multiple projects for a single todo
and also accessibility is improved as earlier the checkboxes were inside a
div, so were not accessible directly using tabbing.