Bug Fix: Fix bug in functioning of delete project button
Event listeners inside for loop were used for capturing clicks on all
delete project buttons, but the problem lied on the fact that current
index of for loop was used to select the project from projects array.
When any of projects were deleted, the projects array was modified but
the index of for loop was referencing the original projects array when
the for loop was set up. This was causing problems. So, in the solution,
index was rather calculated using the clicked button's position in DOM.