Bug Fix: Update projects class when a project is deleted
When projects are added to DOM, each of them is assigned a class according
to their indices. Now, when an project is deleted, the projects following
this project still have clasess according to previous indices, but now their
current indices in projects array are changed. This was causing problems,
hence needed a fix.